Fixed JIT calling wrong debug handler
When a function is compiled by the JIT it keeps the address to the
function which was set as dbreak callback when it compiled. If the
function changed later on, it would still try to call the old function.

Wrap the actual dbreak function call in a GlobalDebugBreak function,
which calls the correct one.